+++ src/main/java/com/example/ImageGalery/Service/ImagenService.java
@@ -3,7 +3,10 @@ package com.example.ImageGalery.Service;
 import com.example.ImageGalery.Model.Imagen;
 import com.example.ImageGalery.Repository.IImagenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -33,9 +36,15 @@ public class ImagenService implements IImagenService {
 
     @Override
     public void editarImagen(Long id, Imagen imagen) {
+        String usernameAuth = SecurityContextHolder.getContext().getAuthentication().getName();
         Imagen imagenExiste = imagenRepository.findById(id).orElse(null);
 
         if (imagenExiste != null){
+
+            if (!imagenExiste.getUsuario().getNombreUsuario().equals(usernameAuth)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN,"No tienes permiso");
+            }
+
             if (imagen.getDescripcion() != null){
                 imagenExiste.setDescripcion(imagen.getDescripcion());
             }
@@ -51,6 +60,13 @@ public class ImagenService implements IImagenService {
 
     @Override
     public void eliminarImagen(Long id) {
+        String usernameAuth = SecurityContextHolder.getContext().getAuthentication().getName();
+        Imagen imagenExiste = imagenRepository.findById(id).orElseThrow(() -> new RuntimeException("Imagen no encontrada"));
+
+        if(!imagenExiste.getUsuario().getNombreUsuario().equals(usernameAuth)){
+            throw new RuntimeException("No tienes permiso");
+        }
+
         imagenRepository.deleteById(id);
     }
